etl/requirements.txt
psycopg2-binary==2.9.9
pandas==2.2.2
requests==2.32.3
python-dotenv==1.0.1
tenacity==8.3.0
